changed the structure of the program
