--- datastructure/trajectory_structure.cpp
@@ -1,2 +1,0 @@
-
-
