+++ src/main.cpp
@@ -4,16 +4,16 @@
 #include "trajectory_data_handling/trajectory_file_manager.h"
 
 int main() {
-    //        reset_database();
-//    create_database();
+//    reset_all_data();
+    create_database();
 //    create_rtree_table();
-//    load_tdrive_dataset();
-//    insert_trajectories_into_trajectory_table();
+    load_tdrive_dataset();
+    insert_trajectories_into_trajectory_table();
 //    load_geolife_dataset();
 //    load_database_into_datastructure();
 //    print_trajectories();
-//        load_trajectories_into_rtree();
-//        reset_database();
+//    load_trajectories_into_rtree();
+//    reset_all_data();
 
     return 0;
 }
