+++ src/trajectory_data_handling/sqlite_querying.h
