+++ src/trajectory_data_handling/trajectory_file_manager.h
