--- tests/trajectory_test.cpp
@@ -1,7 +1,0 @@
-#include "doctest.h"
-#include "../datastructure/trajectory.h"
-
-// eksempel test her
-TEST_CASE("trajectory_handler tests") {
-    CHECK(returntwo() == 2);
-}
